[SMD-398] move historical Failures to a new module
